+++ test/test.carbone.js
@@ -2245,8 +2245,8 @@ describe('Carbone', function () {
         var _elapsed = (_end.getTime() - _start.getTime())/_nbExecuted; // time in milliseconds
         console.log('\n\n Basic rendering: Time Elapsed : '+_elapsed + ' ms per file for '+_nbExecuted+' attempts (usally around 10ms)\n\n\n');
         for (var i = 0; i < _results.length; i++) {
-          var _buf = new Buffer(_results[i]);
-          assert.equal((_buf.slice(0, 2).toString() === 'PK'), true);
+          assert.equal(Buffer.isBuffer(_results[i]), true);
+          assert.equal((_results[i].slice(0, 2).toString() === 'PK'), true);
         }
         assert.equal((_elapsed < 200), true);
         done();
@@ -2584,15 +2584,12 @@ describe('Carbone', function () {
       };
       carbone.render('test_word_render_A.docx', data, {convertTo : 'pdf'}, function (err, result) {
         assert.equal(err, null);
-        var buf = new Buffer(result);
-        assert.equal(buf.slice(0, 4).toString(), '%PDF');
-        var bufPDF = new Buffer(buf.length);
-        fs.open(_pdfResultPath, 'r', function (status, fd) {
-          fs.read(fd, bufPDF, 0, buf.length, 0, function (err, bytesRead, buffer) {
-            assert.equal(buf.slice(0, 4).toString(), '%PDF');
-            assert.equal(buf.slice(8, 50).toString(), buffer.slice(8, 50).toString());
-            done();
-          });
+        assert.equal(result.slice(0, 4).toString(), '%PDF');
+        fs.readFile(_pdfResultPath, function (err, expected) {
+          assert.equal(err+'', 'null');
+          assert.equal(result.slice(0, 4).toString(), '%PDF');
+          assert.equal(result.slice(8, 50).toString(), expected.slice(8, 50).toString());
+          done();
         });
       });
     });
@@ -2605,8 +2602,7 @@ describe('Carbone', function () {
       };
       carbone.render(path.resolve('./test/datasets/test_word_render_A.docx'), data, {convertTo : 'pdf'}, function (err, result) {
         assert.equal(err, null);
-        var buf = new Buffer(result);
-        assert.equal(buf.slice(0, 4).toString(), '%PDF');
+        assert.equal(result.slice(0, 4).toString(), '%PDF');
         done();
       });
     });
@@ -2655,8 +2651,8 @@ describe('Carbone', function () {
           var _elapsed = (_end.getTime() - _start.getTime())/_nbExecuted; // time in milliseconds
           console.log('\n\n Conversion to PDF Time Elapsed : '+_elapsed + ' ms per pdf for '+_nbExecuted+' conversions (usally around 65ms) \n\n\n');
           for (var i = 0; i < _results.length; i++) {
-            var _buf = new Buffer(_results[i]);
-            assert.equal(_buf.slice(0, 4).toString(), '%PDF');
+            assert.equal(Buffer.isBuffer(_results[i]), true);
+            assert.equal(_results[i].slice(0, 4).toString(), '%PDF');
           }
           // assert.equal((_elapsed < 200), true);
           done();