+++ CHANGELOG.md
@@ -161,6 +161,18 @@
 
     example: `{d.myImage:imageFit(contain)}` or `{d.myImage:imageFit(fill)}`
 
+  - Added Libreoffice export filters for PDF rendering. To apply filters to a PDF, it is possible to assign an object to `convertTo` with `formatName:'pdf'` and  `formatOptions` an object with the specified filters. The filter list is available on the documentation. Here is an example of a PDF with a password and a watermark:
+    ```js
+      options = {
+        convertTo: {
+          formatName: 'pdf',
+          formatOptions: {
+            EncryptFile          : true,
+            DocumentOpenPassword : 'QWERTY1234',
+            Watermark            : 'Watermark Carbone'
+          }
+        }
+      }